Fixed send tick issues

+++ Source/Supremacy/Public/BPFL_Helpers.h
@@ -22,7 +22,7 @@ public:
 
 	/** Pack WarMachine Move Update  */
 	UFUNCTION(BlueprintPure, Category = "Network", DisplayName = "Pack WarMachine Move Update", meta=(Keywords = "Net Message Bytes Byte Array Binary"))
-	static void PackWarMachineMoveUpdate(const uint8 Number, const int X, const int Y, const float Yaw, TArray<uint8> &Bytes);
+	static void PackWarMachineMoveUpdate(const uint8 Number, const int X, const int Y, const int Yaw, TArray<uint8> &Bytes);
 	
 	/** Converts a String to an Array of bytes */
 	UFUNCTION(BlueprintPure, Category = "Helpers", DisplayName = "Convert String To Bytes")
